+++ tests/files/type/test_json.py
@@ -3,7 +3,7 @@ import tempfile
 
 import pandas as pd
 
-from astro.files.type import JSONFileType
+from astro.files.types import JSONFileType
 
 sample_file = pathlib.Path(
     pathlib.Path(__file__).parent.parent.parent, "data/sample.json"
